[Update] Update the page.js of forgot password
